@@ -1,14 +1,16 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 22
-    buildToolsVersion "21.1.2"
+    compileSdkVersion 32
+//    buildToolsVersion "21.1.2"
+    buildToolsVersion "30.0.3"
+
     defaultConfig {
         applicationId "info.quangminhdang.linkintercept"
         minSdkVersion 15
-        targetSdkVersion 22
+        targetSdkVersion 32
         versionCode 3
-        versionName '1.2'
+        versionName "1.2"
     }
     buildTypes {
         release {
@@ -16,12 +18,12 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-    productFlavors {
-    }
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    testCompile 'junit:junit:4.12'
-    compile 'com.android.support:appcompat-v7:22.2.1'
+    def appcompat_version = "1.4.2"
+
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    testImplementation 'junit:junit:4.12'
+    implementation "androidx.appcompat:appcompat:$appcompat_version"
 }
